Batch API no longer inherits from API
